Added support for IDIV and IDIVK
Fiu now uses buffer lib for reading the bytecode for luau_deserialize
instead of string.unpack

Removed luau_newmodule, luau_newproto
Fiu now manually calls table.create with the required size, since the
size of the lists are given in the bytecode.
This means these two functions will not be included in the returned
module either.

luau_deserialize will first gather all the information about module
and protos before creating a dictionary to store them.

Removed extended_op_list which was used to check if an opcode has aux
luau_deserialize now uses the already acquired op_info to check if
info[3] is true to find out if the opcode has aux data or not.

Removed most table.insert calls, since they were mostly being used to
populate an array in a for loop with the index being known.